switch to the built-in ed25519 library
Also correctly parse private keys that don't have a redundant public key
appended. This commit doesn't change the *serialization* format for backwards
compatibility but we can now do that in the future (once most users are using
this version).

This *does* remove the `ToCurve25519` function as it's not in the standard
library. However:

1. Nobody's using it.
2. Users can always extract the serialized key and do this themselves (by
calling `Raw()`).